[FEAT] 영화리스트 함수 이동
