+++ src/front/src/ticketing/booking/Calender.scss
@@ -52,7 +52,7 @@ svg {
     height: 50px;
     font-size: 15px;
     line-height: 45px;
-    font-style: italic;
+    //font-style: italic;
 }
 }
 .body {
