+++ src/front/src/ticketing/booking/Ticketing.css
@@ -79,13 +79,19 @@
     margin: 0 auto;
 }
 
-.selectmv{
+.newMv{
+    border: 1px solid gray;
     width: 20%;
     height: 450px;
-    border: 1px solid gray;
-    font-size: 15px;
     margin-top: 2%;
-    float: left;
+}
+
+.selectmv{
+    width: 100%;
+    height: 390px;
+    font-size: 15px;
+    border-top: 1px solid lightgrey;
+    margin-top: 6.3%;
     overflow: scroll;
     overflow-x: hidden;
 }
